Fix: 이미지 적용 통일을 위해 svgr 적용과 폴더 구조 변경
